+++ crates/core/src/lib.rs
@@ -4,15 +4,17 @@
 
 pub mod backup;
 pub mod config;
-pub mod wallet;
-pub mod logger;
-pub mod node;
 pub mod error;
 pub mod fs;
-pub mod theme;
+pub mod logger;
 pub mod network;
+pub mod node;
+pub mod style;
+pub mod theme;
 #[cfg(feature = "wgpu")]
 pub mod utility;
+pub mod wallet;
+pub mod widget;
 
 #[macro_use]
 extern crate lazy_static;
@@ -21,5 +23,5 @@ extern crate lazy_static;
 extern crate log;
 
 // Re-exports
-pub use grin_util::logger::{LoggingConfig, LogEntry};
 pub use grin_core::consensus::GRIN_BASE;
+pub use grin_util::logger::{LogEntry, LoggingConfig};
