--- crates/widgets/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "grin-gui-widgets"
-description = "Widget library for Grin Core GUI"
-version = "0.1.0-alpha.6"
-authors = ["Yeastplume", "Casper Rogild Storm", "tarkah <admin@tarkah.dev>"]
-license = "GPL-3.0"
-homepage = "https://github.com/mimblewimble/grin-gui"
-repository = "https://github.com/mimblewimble/grin-gui"
-edition = "2018"
-
-[features]
-default = ["wgpu"]
-wgpu = ["iced_wgpu"]
-#opengl = ["iced_glow"]
-
-[dependencies]
-iced = { version = "0.12", features = ["advanced", "lazy", "tokio"] }
-iced_graphics = { version = "0.12" }
-iced_core = { version = "0.12" }
-iced_wgpu = { version = "0.12", optional = true }
-iced_style = "0.12"
