--- crates/widgets/src/lib.rs
@@ -1,9 +1,0 @@
-#![allow(dead_code)]
-#![allow(unused_variables)]
-
-pub mod style;
-pub mod widget;
-
-pub use widget::header;
-pub use widget::table_row;
-
